requirements.txt
torch>=2.5.1
torchvision>=0.20.1
transformers==4.46.3
tokenizers==0.20.3
PyMuPDF
img2pdf
einops
easydict
addict
Pillow
numpy
fastapi
uvicorn
httpx
python-dotenv
python-multipart
